crewai/requirements.txt
crewai
langchain-openai
python-dotenv
litellm
fastapi
apscheduler
email-validator
fastapi_sso
